requirements.txt
annotated-types==0.5.0
alembic==1.11.1
beautifulsoup4==4.12.2
bs4==0.0.1
coverage==7.2.5
fastapi==0.100.0
load-dotenv==0.1.0
pandas==2.0.3
psycopg2-binary==2.9.6
pytest==7.2.2
pytest-cov==4.0.0
pytest-md-report==0.3.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-minifier==2.8.1
soupsieve==2.4.1
SQLAlchemy==2.0.18
stackexplain==1.5.0
tabledata==1.3.1
tabulate==0.9.0
urequests==0.1.2
urllib3==2.0.3
uvicorn==0.22.0
virtualenv==20.19.0